// YQ Management/backend/src/communication/events/communication-events.enum.ts
export enum CommunicationEvent {
  USER_REGISTERED = 'USER_REGISTERED',
  LOGIN_OTP_REQUESTED = 'LOGIN_OTP_REQUESTED',
  SIGNUP_OTP_REQUESTED = 'SIGNUP_OTP_REQUESTED',
  PASSWORD_RESET_REQUESTED = 'PASSWORD_RESET_REQUESTED',
  WORKSPACE_INVITED = 'WORKSPACE_INVITED',
  CUSTOMER_VERIFIED = 'CUSTOMER_VERIFIED',
  QUEUE_JOINED = 'QUEUE_JOINED',
  POSITION_CHANGED = 'POSITION_CHANGED',
  NOW_SERVING = 'NOW_SERVING',
  QUEUE_DELAYED = 'QUEUE_DELAYED',
  QUEUE_COMPLETED = 'QUEUE_COMPLETED',
  QUEUE_CANCELLED = 'QUEUE_CANCELLED',
  FEEDBACK_REQUESTED = 'FEEDBACK_REQUESTED',
  CHECK_IN = 'CHECK_IN',
  TOKEN_TRANSFERRED = 'TOKEN_TRANSFERRED',
  BILLING_PAYMENT_SUCCESS = 'BILLING_PAYMENT_SUCCESS',
  BILLING_PAYMENT_FAILED = 'BILLING_PAYMENT_FAILED',
  BILLING_TRIAL_ENDING = 'BILLING_TRIAL_ENDING',
  BILLING_SUBSCRIPTION_RENEWED = 'BILLING_SUBSCRIPTION_RENEWED',
  BILLING_SUBSCRIPTION_CANCELLED = 'BILLING_SUBSCRIPTION_CANCELLED',
  BILLING_SUBSCRIPTION_EXPIRED = 'BILLING_SUBSCRIPTION_EXPIRED',
  BILLING_PAYMENT_REMINDER = 'BILLING_PAYMENT_REMINDER',
  MARKETING_WELCOME = 'MARKETING_WELCOME',
}